refactor(graph): dedup/union 不変条件を src/graph/canonical.ts に一元化 (refs #164)
- builder.ts のエッジ dedup (source|target|kind キー, keep-first) +
  provenance sort-union + 決定的ソート (raw codeunit 比較, INV-L4) を
  dedupEdges / sortNodesById として canonical.ts へ移動
- rename-lock.ts の unionDeps / sortDeps / DepRef を canonical.ts へ
  移動し、lock.ts は canonical.ts から import (「mirrors builder.ts」
  相互参照コメントによる暗黙の合意を import 構造に置換)
- sortUniqueProvenances / codeunitCompare を共有語彙として公開。
  #163 の finalizeDeterministic パスがそのまま呼べる粒度
- 挙動は byte-identical: dedup キー・union 意味論・ソート順すべて不変
